#32 refactor: 새로운 CropsInfoRepository 구현
